wakeup: Create fake debugfs folder solely for wakeup_sources
Userspace applications and other things expect the wakeup_sources path
to be /sys/kernel/debug/wakeup_sources, which won't be present when
debugfs is disabled.

To satisfy these userspace components when debugfs is disabled, keep
wakeup_sources in the same location that it's usually expected to be in by
creating a fake debugfs directory.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>