msm: camera_v2: conditionality eliminate drop_reconfig flag
The flag is set to notify userspace driver, and expected to be clear
there. However, our userspace driver isn't able to handle this, as a
result the camera stopped working once it's set because an empty frame
is always returned.

This conditionality reverts f4c6a3a21b46
and ecd74203a36d

Change-Id: I598d8ffc5f2d8eb1fe387a5210e0ef97d2e3badd
Signed-off-by: Hachiman270 <hachiman270@gmail.com>
Signed-off-by: zeta96 [ for santoni ]